# alternative fix for mul!