@dkarrasch 2020-01-16
I have wanted to move short-circuiting to the runtime code, but couldn't figure out how to do that with minimal invasion. This is a very elegant way to do that.

I wonder one thing, though: we take care of whether `a` is one, and avoid multiplication if it is. We don't do that with `b`, and haven't done that earlier. Is that because this is a special, yet not default case? Would it make sense/be worth it to add this logic here? It would add a second layer of branching, and I guess the one layer you added is responsible for the little slowdown.

Anyway, I think the fact that this handles 5-arg mul! nicely is a great achievement.